requirements.txt
imaplib2==3.6
email-validator==2.0.0
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1